Kernel/xHCI: Move register definitions to separate header
The xHCIController.h header is currently quite big.
